Minor updates in dns server
